Air _NEUTRAL is not idle with Izanami float as a special case (#30)
* Air _NEUTRAL is not idle with Izanami float as a special case

* Air Barrier is considered not idle

* fixed CharIndex::Bang typo

* Refactored PlayerExtendedData into its own files

* Izanami float time is a SLOT

---------
