+++ src/Game/CharData.h
@@ -138,7 +138,6 @@ to double check, just note.
 	int32_t moveCounterHitAirPushbackY; //0x06D4
 	char pad_06D8[760]; //0x06D8
 
-
 	int32_t previousHP; //0x09D0
 	int32_t currentHP; //0x09D4
 	int32_t maxHP; //0x09D8
@@ -209,10 +208,11 @@ to double check, just note.
 	char pad_5AFC[8]; //0x5AFC
 	int32_t barrier; //0x5B04
 	//char pad_5B08[102072]; //0x5b08
-	char pad_5B08[101524]; //0x5b08
+	char pad_5B08[10364]; //0x5b08
 
 
-
+	int32_t SLOT_unknown1; //0x8384 // Izanami float (0 or 1), possibly other stuff
+	char pad_8388[91156]; //0x8388
 
 
 	/*input buffers*/
